Gridのイテレータなどを実装 modified:   coord.h modified:   grid.h modified:   rl.cpp modified:   size.cpp modified:   size.h

+++ size.cpp
@@ -4,8 +4,8 @@
 namespace rl {
 
 void Size::forEach(std::function<void(const Coord &)> f) const {
-  for (int y = 0; y < height_; y++) {
-    for (int x = 0; x < width_; x++) {
+  for (short y = 0; y < height_; y++) {
+    for (short x = 0; x < width_; x++) {
       f(Coord(x, y));
     }
   }
